Fix a warning in debug mode, and remove useless define guards

+++ libdynamohack/include/sp_lev.h
@@ -211,15 +211,15 @@ enum opcode_defs {
 
 #define SP_MONST_CLASS(l) ((l) & 0xff)
 #define SP_MONST_PM(l)	  (((l) >> 8) & 0xffff)
-#define SP_MONST_PACK(m,c) (((m) << 8) + ((char)(c)))
+#define SP_MONST_PACK(m,c) ((((unsigned int)m) << 8) + ((char)(c)))
 
 #define SP_OBJ_CLASS(l)	  ((l) & 0xff)
 #define SP_OBJ_TYP(l)	  (((l) >> 8) & 0xffff)
-#define SP_OBJ_PACK(o,c)  (((o) << 8) + ((char)(c)))
+#define SP_OBJ_PACK(o,c)  ((((unsigned int)o) << 8) + ((char)(c)))
 
 #define SP_MAPCHAR_TYP(l) ((l) & 0xff)
 #define SP_MAPCHAR_LIT(l) (((l) >> 8) & 0xff)
-#define SP_MAPCHAR_PACK(typ,lit) (((lit) << 8) + ((char)(typ)))
+#define SP_MAPCHAR_PACK(typ,lit) ((((unsigned int)lit) << 8) + ((char)(typ)))
 
 struct opvar {
 	xchar spovartyp; /* one of SPOVAR_foo */
